feat: userRole identify; code optimize

+++ src/main/java/com/cxy/usercenter/controller/UserController.java
@@ -10,7 +10,12 @@ import org.springframework.web.bind.annotation.*;
 
 import javax.annotation.Resource;
 import javax.servlet.http.HttpServletRequest;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
+
+import static com.cxy.usercenter.constant.UserConstant.ADMIN_ROLE;
+import static com.cxy.usercenter.constant.UserConstant.USER_LOGIN_STATE;
 
 /**
  * 用户接口
@@ -58,21 +63,36 @@ public class UserController {
     }
 
     @GetMapping("/search")
-    public List<User> searchUsers(String username) {
+    public List<User> searchUsers(String username, HttpServletRequest request) {
+        //仅管理员可查询
+        if (!isAdmin(request)) {
+            return new ArrayList<>();
+        }
         QueryWrapper<User> queryWrapper = new QueryWrapper<>();
         if (StringUtils.isNotBlank(username)) {
             queryWrapper.like("username", username);
         }
-        return userService.list(queryWrapper);
-//        return userService.searchUsers(username);
+        List<User> userList = userService.list(queryWrapper);
+        return userList.stream().map(user -> userService.getSafetyUser(user)).collect(Collectors.toList());
     }
 
     @PostMapping("/delete")
-    public boolean deleteUser(@RequestBody long id) {
+    public boolean deleteUser(@RequestBody long id, HttpServletRequest request) {
+        //仅管理员可查询
+        if (!isAdmin(request)) {
+            return false;
+        }
         if (id <= 0) {
             return false;
         }
         // 逻辑删除
         return userService.removeById(id);
+    }
+
+    private boolean isAdmin(HttpServletRequest request) {
+        //仅管理员可查询
+        Object userObj = request.getSession().getAttribute(USER_LOGIN_STATE);
+        User user = (User)userObj;
+        return user != null && user.getUserRole() == ADMIN_ROLE;
     }
 }
